Physics Engine Improvements
major physics engine overhaul

the three.js inbuilt raycaster is too slow (it attempts to check every polygon of every instance??) so now we are using our own collision system which stores voxels in an infinite grid as a vector field with two weights.

This new method is more than 10 times faster! (~45ms per ray to ~4ms) and is a superior alternative to threeBVH or other raycasting tools.